backend/requirements.txt
argparse
pandas
PyGithub>=2.3
dotenv
openpyxl